asetypes/convert.go: Combine ConvertValue-methods and process null-types
